ai-server/pyproject.toml
[project]
name = "ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "apify-client>=2.5.0",
    "chromadb>=1.5.1",
    "crewai[tools]>=0.95.0",
    "fastapi[standard]>=0.129.0",
    "firecrawl-py>=4.18.1",
    "httpx>=0.28.1",
    "langchain>=0.3",
    "langchain-chroma>=0.2",
    "langchain-openai>=0.3",
    "loguru>=0.7.3",
    "openai>=2.24.0",
    "prisma>=0.15.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "sqlmodel>=0.0.35",
    "xai-sdk>=1.8.0",
]
